+++ backend/chainlit/slack/app.py
@@ -343,7 +343,15 @@ async def handle_message(message, say):
         )
     )
     thread_name = f"{user.identifier} Slack DM {datetime.today().strftime('%Y-%m-%d')}"
-    await process_slack_message(message, say, thread_name, True)
+    ts = message.get("thread_ts", message["ts"])
+    await process_slack_message(
+        event=message,
+        say=say,
+        thread_id=thread_id,
+        thread_name=thread_name,
+        bind_thread_to_user=True,
+        thread_ts=ts,
+    )
 
 
 @slack_app.block_action("thumbdown")